requirements.txt
aiodns==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
hyperframe==6.0.1
idna==3.4
loguru==0.7.2
lxml==4.9.3
multidict==6.0.4
nonebot2==2.1.1
packaging==23.2
pycares==4.4.0
pycparser==2.21
pydantic==1.10.13
pygtrie==2.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==11.0.3
win32-setctime==1.1.0
yarl==1.9.2
pillow
openai~=0.27.10
pycryptodome==3.19.0
requests~=2.31.0
nonebot_plugin_apscheduler==0.3.0
wechatpy==1.8.18
nonebot-adapter-onebot===2.3.1

